Feat: 휴게시간 필드 추가
- 휴게시간 필드 추가에 따른 파싱 및 재모델링 (구조변경)
- 사용하지 않는 기존 파일 삭제 및 폴더 재구조화
- 폴더 재구조화에 따른 컴파일 에러 해결 (별도의 Extension으로 관리하도록)
