+++ CheckEat-User/Serivce/Map/SearchByVeganLevelService.swift
